+++ oneflow/core/kernel/affine_channel_kernel.h
@@ -30,17 +30,17 @@ class AffineChannelKernel final : public KernelIfWithModel<device_type, T>,
 template<DeviceType device_type, typename T>
 class AffineChannelKernelUtil final {
  public:
-  static void Forward(DeviceCtx* ctx, const int64_t elem_cnt, const int32_t channel_dim,
-                      const int64_t channel_stride, const T* in, const T* scale, const T* bias,
+  static void Forward(DeviceCtx* ctx, const int32_t elem_cnt, const int32_t channel_dim,
+                      const int32_t channel_stride, const T* in, const T* scale, const T* bias,
                       T* out);
-  static void BackwardInDiff(DeviceCtx* ctx, const int64_t elem_cnt, const int32_t channel_dim,
-                             const int64_t channel_stride, const T* out_diff, const T* scale,
+  static void BackwardInDiff(DeviceCtx* ctx, const int32_t elem_cnt, const int32_t channel_dim,
+                             const int32_t channel_stride, const T* out_diff, const T* scale,
                              T* in_diff);
-  static void BackwardScaleBiasDiff(DeviceCtx* ctx, const int64_t elem_cnt,
-                                    const int32_t channel_dim, const int64_t channel_stride,
+  static void BackwardScaleBiasDiff(DeviceCtx* ctx, const int32_t elem_cnt,
+                                    const int32_t channel_dim, const int32_t channel_stride,
                                     const T* in, const T* out_diff, T* scale_diff, T* bias_diff);
-  static void BackwardScaleDiff(DeviceCtx* ctx, const int64_t elem_cnt, const int32_t channel_dim,
-                                const int64_t channel_stride, const T* in, const T* out_diff,
+  static void BackwardScaleDiff(DeviceCtx* ctx, const int32_t elem_cnt, const int32_t channel_dim,
+                                const int32_t channel_stride, const T* in, const T* out_diff,
                                 T* scale_diff);
 };
